pyproject.toml
[project]
name = "phanloaivithuoc-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "aiortc>=1.14.0",
    "easyrtc>=0.0.2",
    "fastapi>=0.120.0",
    "fastrtc>=0.0.33",
    "ffmpeg>=1.4",
    "motor>=3.7.1",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "python-snap7>=2.0.2",
    "ultralytics>=8.3.221",
    "uvicorn[standard]>=0.38.0",
]
